+++ packages/autorest.go/src/generator/generator.ts
@@ -60,6 +60,8 @@ export async function generateCode(host: AutorestExtensionHost) {
       });
     }
 
+    const factoryGatherAllParams = await session.getValue('factory-gather-all-params', false);
+    session.model.options.factoryGatherAllParams = factoryGatherAllParams;
     const clientFactory = await generateClientFactory(session.model);
     if (clientFactory.length > 0) {
       host.writeFile({